// src/app/users/model/user.ts
// {
//  "id":"035fccc8-956e-4f71-b68d-9f97d752cb73",
//  "firstName":"Rashad",
//  "lastName":"Weiden",
//  "email":"dernemploys@hotmail.co.uk",
//  "phone":"6202457020",
//  "city":"Chisinau"
// }

export interface User {
  id?: string;
  firstName: string;
  lastName: string;
  email: string;
  phone: string;
  city: string;
}